+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -229,6 +229,54 @@ export interface Site {
   updateOrganizationId: (organizationId: string) => Site;
 }
 
+/**
+ * Represents a site candidate.
+ */
+export interface SiteCandidate {
+  /**
+   * Retrieves the base URL of the site candidate.
+   * @returns {string} The base URL.
+   */
+  getBaseURL: () => string;
+
+  /**
+   * Retrieves the site id of the site candidate.
+   * Only set after APPROVED state
+   * @returns {string} site id
+   */
+  getSiteId: () => string;
+
+  /**
+   * Retrieves the source of the site candidate.
+   * @returns {string} The source
+   */
+  getSource: () => string;
+
+  /**
+   * Retrieves the status of the site candidate.
+   * @returns {string} The status
+   */
+  getStatus: () => string;
+
+  /**
+   * Retrieves the creation timestamp of the site candidate.
+   * @returns {string} The creation timestamp.
+   */
+  getCreatedAt: () => string;
+
+  /**
+   * Retrieves the last update timestamp of the site candidate.
+   * @returns {string} The last update timestamp.
+   */
+  getUpdatedAt: () => string;
+
+  /**
+   * Retrieves the slack id of the person who last updated the site candidate.
+   * @returns {string} The last update timestamp.
+   */
+  getUpdatedBy: () => string;
+}
+
 export interface Organization {
   /**
    * Retrieves the ID of the site.
@@ -353,6 +401,11 @@ export interface DataAccess {
   removeOrganization: (
       organizationId: string,
   ) => Promise<void>;
+
+  // site candidate functions
+  upsertSiteCandidate: (siteCandidateDate: object) => Promise<SiteCandidate>;
+  siteCandidateExists: (baseURL: string) => Promise<boolean>;
+  updateSiteCandidate: (siteCandidate: SiteCandidate) => Promise<SiteCandidate>;
 }
 
 interface DataAccessConfig {
@@ -360,6 +413,7 @@ interface DataAccessConfig {
   tableNameLatestAudits: string;
   tableNameOrganizations: string,
   tableNameSites: string;
+  tableNameSiteCandidates: string;
   indexNameAllSites: string;
   indexNameAllSitesOrganizations: string,
   indexNameAllSitesByDeliveryType: string;