+++ packages/spacecat-shared-slack-client/src/index.js
@@ -9,77 +9,18 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { WebClient } from '@slack/web-api';
-import { hasText, isObject } from '@adobe/spacecat-shared-utils';
 
-const ENV_PREFIX = 'SLACK_TOKEN_';
+import BaseSlackClient from './clients/base-slack-client.js';
+import ElevatedSlackClient, { SLACK_STATUSES } from './clients/elevated-slack-client.js';
 
-export const SLACK_TARGETS = {
-  ADOBE_INTERNAL: 'ADOBE_INTERNAL',
-  ADOBE_EXTERNAL: 'ADOBE_EXTERNAL',
+const SLACK_TARGETS = {
+  WORKSPACE_INTERNAL: 'WORKSPACE_INTERNAL',
+  WORKSPACE_EXTERNAL: 'WORKSPACE_EXTERNAL',
 };
 
-function getEnvironmentVariableNameForTarget(target) {
-  return `${ENV_PREFIX}${target}`;
-}
-
-export class SlackClient {
-  static createFrom(context, target) {
-    const { log } = context;
-
-    if (!hasText(target)) {
-      throw new Error('Missing target for the Slack Client');
-    }
-
-    if (!isObject(context.slackClients)) {
-      context.slackClients = {};
-    }
-
-    if (context.slackClients[target]) {
-      return context.slackClients[target];
-    }
-
-    const token = context.env[getEnvironmentVariableNameForTarget(target)];
-    if (!hasText(token)) {
-      throw new Error(`No slack token set for ${target}`);
-    }
-
-    context.slackClients[target] = new SlackClient(token, log);
-    return context.slackClients[target];
-  }
-
-  constructor(token, log) {
-    this.client = new WebClient(token);
-    this.log = log;
-  }
-
-  async #apiCall(method, message) {
-    try {
-      return await this.client.apiCall(method, message);
-    } catch (e) {
-      this.log.error(`Failed to perform slack api call: ${method}`, e);
-      throw e;
-    }
-  }
-
-  async postMessage(message) {
-    const result = await this.#apiCall('chat.postMessage', message);
-    return {
-      channelId: result.channel,
-      threadId: result.ts,
-    };
-  }
-
-  async fileUpload(file) {
-    const result = await this.#apiCall('files.uploadV2', file);
-
-    if (result.files.length === 0) {
-      throw new Error(`File upload was unsuccessful. Filename was "${file.filename}"`);
-    }
-
-    return {
-      fileUrl: result.files[0].url_private,
-      channels: result.files[0].channels,
-    };
-  }
-}
+export {
+  SLACK_STATUSES,
+  SLACK_TARGETS,
+  BaseSlackClient,
+  ElevatedSlackClient,
+};