+++ packages/spacecat-shared-slack-client/CHANGELOG.md
@@ -1,3 +1,25 @@
+# [@adobe/spacecat-shared-slack-client-v1.1.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-slack-client-v1.1.1...@adobe/spacecat-shared-slack-client-v1.1.2) (2024-02-05)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#129](https://github.com/adobe/spacecat-shared/issues/129)) ([22b6c90](https://github.com/adobe/spacecat-shared/commit/22b6c90aea25783116165fa1dacb08c631c792bc))
+* **revert:** remove slack targets type ([dbd64d4](https://github.com/adobe/spacecat-shared/commit/dbd64d46ec64d6936558e83d60756cddb2fce446))
+
+# [@adobe/spacecat-shared-slack-client-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-slack-client-v1.1.0...@adobe/spacecat-shared-slack-client-v1.1.1) (2024-02-02)
+
+
+### Bug Fixes
+
+* add slack targets type ([#127](https://github.com/adobe/spacecat-shared/issues/127)) ([537c01b](https://github.com/adobe/spacecat-shared/commit/537c01be6d8bca992bf21d68abc7f8e2930ab169))
+
+# [@adobe/spacecat-shared-slack-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-slack-client-v1.0.1...@adobe/spacecat-shared-slack-client-v1.1.0) (2024-02-02)
+
+
+### Features
+
+* slack channel management ([#124](https://github.com/adobe/spacecat-shared/issues/124)) ([37aa44c](https://github.com/adobe/spacecat-shared/commit/37aa44cd36e28895766c16021eb76fe1b4d7f071))
+
 # [@adobe/spacecat-shared-slack-client-v1.0.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-slack-client-v1.0.0...@adobe/spacecat-shared-slack-client-v1.0.1) (2024-01-26)