+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@adobe/spacecat-shared-utils-v1.9.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.8.0...@adobe/spacecat-shared-utils-v1.9.0) (2024-02-06)
+
+
+### Features
+
+* introduce AWS S3 wrapper ([#135](https://github.com/adobe/spacecat-shared/issues/135)) ([4ad1bb6](https://github.com/adobe/spacecat-shared/commit/4ad1bb64414086af5d399317697debb6aff0fb74))
+
+# [@adobe/spacecat-shared-utils-v1.8.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.7.4...@adobe/spacecat-shared-utils-v1.8.0) (2024-02-05)
+
+
+### Features
+
+* centralize sqsEventAdapter in spacecat-shared-utils ([#128](https://github.com/adobe/spacecat-shared/issues/128)) ([52ec442](https://github.com/adobe/spacecat-shared/commit/52ec4427d8d2c989b369521d30699e90ee8a83cc))
+
+# [@adobe/spacecat-shared-utils-v1.7.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.7.3...@adobe/spacecat-shared-utils-v1.7.4) (2024-02-01)
+
+
+### Bug Fixes
+
+* secrets resolve ci versions properly ([#66](https://github.com/adobe/spacecat-shared/issues/66)) ([2ddebdd](https://github.com/adobe/spacecat-shared/commit/2ddebddb6c76d633f6be0a5bca3ee3bd150ef474))
+
 # [@adobe/spacecat-shared-utils-v1.7.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.7.2...@adobe/spacecat-shared-utils-v1.7.3) (2024-01-23)